+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 **Added**
 
+- #537 Instrument import interface: Sysmex XT-4000i
+- #536 Instrument import interface: Sysmex XT-1800i
 - #618 When previewing stickers the number of copies to print for each sticker can be modified.
 - #618 The default number of sticker copies can be set and edited in the setup Sticker's tab.